refactor: clean up around scores

+++ core/src/main/java/ai/timefold/solver/core/impl/score/stream/common/inliner/BendableScoreContext.java
@@ -3,7 +3,7 @@ package ai.timefold.solver.core.impl.score.stream.common.inliner;
 import java.util.Arrays;
 import java.util.Objects;
 
-import ai.timefold.solver.core.api.score.buildin.bendable.BendableScore;
+import ai.timefold.solver.core.api.score.BendableScore;
 import ai.timefold.solver.core.impl.score.stream.common.AbstractConstraint;
 
 import org.jspecify.annotations.NullMarked;
@@ -13,11 +13,11 @@ final class BendableScoreContext extends ScoreContext<BendableScore, BendableSco
     private final int hardScoreLevelCount;
     private final int softScoreLevelCount;
     private final int scoreLevel;
-    private final int scoreLevelWeight;
+    private final long scoreLevelWeight;
 
     public BendableScoreContext(BendableScoreInliner parent, AbstractConstraint<?, ?, ?> constraint,
             BendableScore constraintWeight, int hardScoreLevelCount, int softScoreLevelCount, int scoreLevel,
-            int scoreLevelWeight) {
+            long scoreLevelWeight) {
         super(parent, constraint, constraintWeight);
         this.hardScoreLevelCount = hardScoreLevelCount;
         this.softScoreLevelCount = softScoreLevelCount;
@@ -30,7 +30,7 @@ final class BendableScoreContext extends ScoreContext<BendableScore, BendableSco
         this(parent, constraint, constraintWeight, hardScoreLevelCount, softScoreLevelCount, -1, -1);
     }
 
-    public ScoreImpact<BendableScore> changeSoftScoreBy(int matchWeight,
+    public ScoreImpact<BendableScore> changeSoftScoreBy(long matchWeight,
             ConstraintMatchSupplier<BendableScore> constraintMatchSupplier) {
         var softImpact = Math.multiplyExact(scoreLevelWeight, matchWeight);
         inliner.softScores[scoreLevel] = Math.addExact(inliner.softScores[scoreLevel], softImpact);
@@ -38,7 +38,7 @@ final class BendableScoreContext extends ScoreContext<BendableScore, BendableSco
         return possiblyAddConstraintMatch(scoreImpact, constraintMatchSupplier);
     }
 
-    public ScoreImpact<BendableScore> changeHardScoreBy(int matchWeight,
+    public ScoreImpact<BendableScore> changeHardScoreBy(long matchWeight,
             ConstraintMatchSupplier<BendableScore> constraintMatchSupplier) {
         var hardImpact = Math.multiplyExact(scoreLevelWeight, matchWeight);
         inliner.hardScores[scoreLevel] = Math.addExact(inliner.hardScores[scoreLevel], hardImpact);
@@ -46,10 +46,10 @@ final class BendableScoreContext extends ScoreContext<BendableScore, BendableSco
         return possiblyAddConstraintMatch(scoreImpact, constraintMatchSupplier);
     }
 
-    public ScoreImpact<BendableScore> changeScoreBy(int matchWeight,
+    public ScoreImpact<BendableScore> changeScoreBy(long matchWeight,
             ConstraintMatchSupplier<BendableScore> constraintMatchSupplier) {
-        var hardImpacts = new int[hardScoreLevelCount];
-        var softImpacts = new int[softScoreLevelCount];
+        var hardImpacts = new long[hardScoreLevelCount];
+        var softImpacts = new long[softScoreLevelCount];
         for (var hardScoreLevel = 0; hardScoreLevel < hardScoreLevelCount; hardScoreLevel++) {
             var hardImpact = Math.multiplyExact(constraintWeight.hardScore(hardScoreLevel), matchWeight);
             hardImpacts[hardScoreLevel] = hardImpact;
@@ -65,7 +65,8 @@ final class BendableScoreContext extends ScoreContext<BendableScore, BendableSco
     }
 
     @NullMarked
-    private record SingleSoftImpact(BendableScoreContext ctx, int impact) implements ScoreImpact<BendableScore> {
+    private record SingleSoftImpact(BendableScoreContext ctx,
+            long impact) implements ScoreImpact<BendableScore> {
 
         @Override
         public void undo() {
@@ -79,7 +80,8 @@ final class BendableScoreContext extends ScoreContext<BendableScore, BendableSco
     }
 
     @NullMarked
-    private record SingleHardImpact(BendableScoreContext ctx, int impact) implements ScoreImpact<BendableScore> {
+    private record SingleHardImpact(BendableScoreContext ctx,
+            long impact) implements ScoreImpact<BendableScore> {
 
         @Override
         public void undo() {
@@ -93,8 +95,8 @@ final class BendableScoreContext extends ScoreContext<BendableScore, BendableSco
     }
 
     @NullMarked
-    private record ComplexImpact(BendableScoreContext ctx, int[] hardImpacts,
-            int[] softImpacts) implements ScoreImpact<BendableScore> {
+    private record ComplexImpact(BendableScoreContext ctx, long[] hardImpacts,
+            long[] softImpacts) implements ScoreImpact<BendableScore> {
 
         @Override
         public void undo() {
