+++ core/src/main/java/ai/timefold/solver/core/impl/score/stream/common/inliner/HardMediumSoftScoreInliner.java
@@ -2,16 +2,16 @@ package ai.timefold.solver.core.impl.score.stream.common.inliner;
 
 import java.util.Map;
 
-import ai.timefold.solver.core.api.score.buildin.hardmediumsoft.HardMediumSoftScore;
+import ai.timefold.solver.core.api.score.HardMediumSoftScore;
 import ai.timefold.solver.core.api.score.stream.Constraint;
 import ai.timefold.solver.core.impl.score.constraint.ConstraintMatchPolicy;
 import ai.timefold.solver.core.impl.score.stream.common.AbstractConstraint;
 
 final class HardMediumSoftScoreInliner extends AbstractScoreInliner<HardMediumSoftScore> {
 
-    int hardScore;
-    int mediumScore;
-    int softScore;
+    long hardScore;
+    long mediumScore;
+    long softScore;
 
     HardMediumSoftScoreInliner(Map<Constraint, HardMediumSoftScore> constraintWeightMap,
             ConstraintMatchPolicy constraintMatchPolicy) {
