+++ core/src/main/java/ai/timefold/solver/core/impl/score/stream/common/inliner/HardSoftScoreContext.java
@@ -1,6 +1,6 @@
 package ai.timefold.solver.core.impl.score.stream.common.inliner;
 
-import ai.timefold.solver.core.api.score.buildin.hardsoft.HardSoftScore;
+import ai.timefold.solver.core.api.score.HardSoftScore;
 import ai.timefold.solver.core.impl.score.stream.common.AbstractConstraint;
 
 import org.jspecify.annotations.NullMarked;
@@ -12,7 +12,7 @@ final class HardSoftScoreContext extends ScoreContext<HardSoftScore, HardSoftSco
         super(parent, constraint, constraintWeight);
     }
 
-    public ScoreImpact<HardSoftScore> changeSoftScoreBy(int matchWeight,
+    public ScoreImpact<HardSoftScore> changeSoftScoreBy(long matchWeight,
             ConstraintMatchSupplier<HardSoftScore> constraintMatchSupplier) {
         var softImpact = Math.multiplyExact(constraintWeight.softScore(), matchWeight);
         inliner.softScore = Math.addExact(inliner.softScore, softImpact);
@@ -20,7 +20,7 @@ final class HardSoftScoreContext extends ScoreContext<HardSoftScore, HardSoftSco
         return possiblyAddConstraintMatch(scoreImpact, constraintMatchSupplier);
     }
 
-    public ScoreImpact<HardSoftScore> changeHardScoreBy(int matchWeight,
+    public ScoreImpact<HardSoftScore> changeHardScoreBy(long matchWeight,
             ConstraintMatchSupplier<HardSoftScore> constraintMatchSupplier) {
         var hardImpact = Math.multiplyExact(constraintWeight.hardScore(), matchWeight);
         inliner.hardScore = Math.addExact(inliner.hardScore, hardImpact);
@@ -28,7 +28,7 @@ final class HardSoftScoreContext extends ScoreContext<HardSoftScore, HardSoftSco
         return possiblyAddConstraintMatch(scoreImpact, constraintMatchSupplier);
     }
 
-    public ScoreImpact<HardSoftScore> changeScoreBy(int matchWeight,
+    public ScoreImpact<HardSoftScore> changeScoreBy(long matchWeight,
             ConstraintMatchSupplier<HardSoftScore> constraintMatchSupplier) {
         var hardImpact = Math.multiplyExact(constraintWeight.hardScore(), matchWeight);
         var softImpact = Math.multiplyExact(constraintWeight.softScore(), matchWeight);
@@ -39,7 +39,8 @@ final class HardSoftScoreContext extends ScoreContext<HardSoftScore, HardSoftSco
     }
 
     @NullMarked
-    private record SoftImpact(HardSoftScoreInliner inliner, int softImpact) implements ScoreImpact<HardSoftScore> {
+    private record SoftImpact(HardSoftScoreInliner inliner,
+            long softImpact) implements ScoreImpact<HardSoftScore> {
 
         @Override
         public void undo() {
@@ -54,7 +55,8 @@ final class HardSoftScoreContext extends ScoreContext<HardSoftScore, HardSoftSco
     }
 
     @NullMarked
-    private record HardImpact(HardSoftScoreInliner inliner, int hardImpact) implements ScoreImpact<HardSoftScore> {
+    private record HardImpact(HardSoftScoreInliner inliner,
+            long hardImpact) implements ScoreImpact<HardSoftScore> {
 
         @Override
         public void undo() {
@@ -69,8 +71,8 @@ final class HardSoftScoreContext extends ScoreContext<HardSoftScore, HardSoftSco
     }
 
     @NullMarked
-    private record ComplexImpact(HardSoftScoreInliner inliner, int hardImpact,
-            int softImpact) implements ScoreImpact<HardSoftScore> {
+    private record ComplexImpact(HardSoftScoreInliner inliner, long hardImpact,
+            long softImpact) implements ScoreImpact<HardSoftScore> {
 
         @Override
         public void undo() {
